requirements.txt
# Core testing dependencies
pytest>=7.0.0
requests>=2.28.0
jsonschema>=4.0.0

# Performance testing
locust>=2.0.0

# Test reporting (optional)
pytest-html>=3.1.0
pytest-cov>=4.0.0
allure-pytest>=2.15.0
allure-python-commons~=2.15.0

# Development dependencies (optional)
pytest-xdist>=2.5.0  # for parallel test execution
